style: split off list into it's own module
